Move image element-specific validators to ImageElement.tsx
