+++ src/plugin/helpers/validation.ts
@@ -2,7 +2,10 @@ import type { FieldValidationErrors, ValidationError } from "../elementSpec";
 import type { FieldNameToValueMap } from "../fieldViews/helpers";
 import type { FieldDescriptions } from "../types/Element";
 
-type Validator = (fieldValue: unknown, fieldName: string) => ValidationError[];
+export type Validator = (
+  fieldValue: unknown,
+  fieldName: string
+) => ValidationError[];
 
 export const createValidator = (
   fieldValidationMap: Record<string, Validator[]>
@@ -20,82 +23,6 @@ export const createValidator = (
     errors[fieldName] = fieldErrors;
   }
   return errors;
-};
-
-type ImageAsset = {
-  fields: {
-    width: number;
-    height: number;
-  };
-};
-
-const hasOwnProperty = <X extends Record<string, unknown>, Y extends string>(
-  obj: X,
-  prop: Y
-): obj is X & Record<Y, unknown> => {
-  return Object.hasOwnProperty.call(obj, prop);
-};
-
-const isRecord = (value: unknown): value is Record<string, unknown> => {
-  return typeof value === "object" && value !== null;
-};
-
-const isValidImageAsset = (
-  maybeImage: Record<string, unknown>
-): maybeImage is ImageAsset => {
-  return (
-    hasOwnProperty(maybeImage, "fields") &&
-    isRecord(maybeImage.fields) &&
-    hasOwnProperty(maybeImage.fields, "width") &&
-    hasOwnProperty(maybeImage.fields, "height") &&
-    typeof maybeImage.fields.width === "number" &&
-    typeof maybeImage.fields.height === "number"
-  );
-};
-
-const validateAssets = (maybeAssets: unknown[]) => {
-  const assets = maybeAssets.map((asset, i) => {
-    if (!isRecord(asset)) {
-      throw new Error(
-        `[largestAssetMinDimension]: asset ${i} passed to validator was not an object`
-      );
-    }
-    if (!isValidImageAsset(asset)) {
-      throw new Error(
-        `[largestAssetMinDimension]: asset ${i} does not have height and width props that are numbers`
-      );
-    }
-    return asset;
-  });
-  return assets;
-};
-
-export const largestAssetMinDimension = (minSize: number): Validator => (
-  value
-) => {
-  if (typeof value !== "object" || value === null) {
-    throw new Error(
-      `[largestAssetMinDimension]: overall value passed to validator is not an object`
-    );
-  }
-
-  if (isRecord(value) && value.assets && Array.isArray(value.assets)) {
-    const validatedAssets = validateAssets(value.assets);
-    const largestImageAsset = validatedAssets.sort(function (a, b) {
-      return b.fields.width - a.fields.width;
-    })[0];
-
-    const largestDimensionMin = minSize;
-
-    if (
-      largestImageAsset.fields.width < largestDimensionMin &&
-      largestImageAsset.fields.height < largestDimensionMin
-    ) {
-      return ["Warning: Small image, only thumbnail available"];
-    }
-  }
-
-  return [];
 };
 
 export const htmlMaxLength = (
